requirements.txt
asgiref==3.2.5
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
autopep8==1.4.4
backports.csv==1.0.7
boto3==1.9.108
botocore==1.12.108
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.2
channels==2.2.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
daphne==2.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==3.0.4
django-cors-headers==2.2.0
django-excel==0.0.10
django-import-export==1.2.0
django-storages==1.7.1
djangorestframework==3.8.2
docutils==0.14
et-xmlfile==1.0.1
facebook-sdk==3.1.0
geoip2==2.9.0
google-auth==1.6.3
graphviz==0.13.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.17
jdcal==1.4
jmespath==0.9.4
lazy-object-proxy==1.3.1
lml==0.0.9
lxml==4.4.2
maxminddb==1.4.1
mccabe==0.6.1
nexmo==2.3.0
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.1
parsel==1.5.2
pbr==4.0.3
pipenv==2018.11.26
Protego==0.1.15
psycopg2==2.8.4
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyexcel==0.5.13
pyexcel-io==0.5.17
pyexcel-webio==0.1.4
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.1.0
pypiwin32==223
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2018.4
pywin32==224
PyYAML==5.1
queuelib==1.5.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
Scrapy==1.8.0
sendgrid==6.0.5
service-identity==18.1.0
six==1.11.0
sqlparse==0.3.0
stevedore==1.28.0
tablib==0.13.0
texttable==1.6.1
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper-win==1.2.6
w3lib==1.21.0
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
